Merge in IN/integrationpoints from develop to master

* commit 'de232ced4208c6f765553257f4d75b7e1f16484d': (22 commits)
  REL-793750 Update IAPI 1.19.4
  REL-703856 Refactored IntegrationPointService
  REL-761335 Removed locking in JobStopManager
  [REL-774373] StringBuilder implemented to throw MalformedLineException
  REL-703856 Added unit tests
  REL-703856 Added retry using Sync app
  REL-703856 Change the way how IsRetryingErrors works for Sync
  [REL-774373] Tests fixed
  [REL-774373] Review remarks.
  REL-709965 Upgraded jquery and jquery-ui
  [REL-774373] Clean up
  [REL-774373] MalformedLineException test for DelimitedFileParser created
  [REL-774373] TextFieldParserWrapper introduced
  [REL-774373] RdoSynchronization SyncData Tests created
  [REL-774373] Refactorization
  [REL-774373] Cleanup
  [REL-774373] BatchStartingIndex changed to nullable for backward compatibility
  [REL-774373] Item level error created for MalformedLineException
  [REL-774373] Tests updated for BatchStartingIndex
  [REL-774373] BatchStartingIndex implemented
  ...

+++ Source/kCura.IntegrationPoints.FtpProvider.Helpers/Exceptions.cs
@@ -32,7 +32,7 @@ namespace kCura.IntegrationPoints.FtpProvider.Helpers
 
         public class NumberOfColumnsNotEqualToNumberOfDataValuesException : Exception
         {
-            public NumberOfColumnsNotEqualToNumberOfDataValuesException(Int32 lineNumber)
+            public NumberOfColumnsNotEqualToNumberOfDataValuesException(long lineNumber)
                 : base($"The are a different number of data columns than column heading in line number {lineNumber}.")
             { }
             public NumberOfColumnsNotEqualToNumberOfDataValuesException(string message)
